create AccountingService,a dn two interface for testing and mocking dependencies

+++ Backend/src/main/java/com/codecool/backend/controller/dto/NewTransactionDto.java
@@ -1,6 +1,7 @@
 package com.codecool.backend.controller.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record NewTransactionDto(String name, List<Long> categoryIds, int amount) {
+public record NewTransactionDto(String name, List<Long> categoryIds, BigDecimal amount) {
 }
